Scaffold multi-camera session + calibration + triangulation
- Session model (multicam.py): N-camera dir discovery, manifest + glob
  fallback, sync_offset honoured by iter_synchronized_frames
- Calibration data layer (calibration.py): JSON IO + schema validation,
  world-frame check, ChArUco solve stub
- Triangulation (triangulation.py): real projection_matrix / project_points
  / undistort_points / triangulate_views (weighted linear DLT),
  fuse_session_frame as policy stub
- pose-estimation-calibrate CLI with verify/solve/capture subcommands
- --session-dir / --sessions-dir / --calibration flags on main + run
- TypedDicts: CameraCalibration, SessionCalibration, SessionFrame,
  MultiCamPipelineState
- Tech docs: .claude/tech/multicam.md, calibration.md
- Tests: 45 new (calibration 18 + multicam 16 + triangulation 11)

+++ src/pose_estimation/drawing.py
@@ -230,7 +230,7 @@ def _blended_color(color):
     frame drawing path indexes those tables directly instead of re-running
     the multiply/round per limb.
     """
-    return tuple(int(round(c * _LIMB_ALPHA)) for c in color)
+    return tuple(round(c * _LIMB_ALPHA) for c in color)
 
 
 _BODY_BLENDED = {k: _blended_color(v) for k, v in BODY_COLOR_MAP.items()}
@@ -259,7 +259,7 @@ def draw_body_landmarks(img, body_landmarks, body_visibilities, visibility_thres
         points = landmarks[:, :2]
         # Convert to Python lists once — the chain/segment/dot loops below
         # all read by index, and per-iteration numpy → Python int conversion
-        # was a measurable cost at 33 keypoints × multiple call sites.
+        # was a measurable cost at 33 keypoints x multiple call sites.
         pts_list = points.astype(np.int32, copy=False).tolist()
         vis_list = (visibility > visibility_threshold).tolist()
